+++ src/main/java/com/exercise/basic/SoHoc/Bai5.java
@@ -1,4 +1,4 @@
-package com.exercise.basic;
+package com.exercise.basic.SoHoc;
 
 public class Bai5 {
